Found error which made the drivers crazy, still errors in BLDC and statemachine because of missing functions.

+++ Bachelor/Core/Code/SystemMeasurements/ADC_Measurements.c
@@ -27,7 +27,7 @@ float H2 = 0;
 float BatteryVol;
 float MDCVol = 0;
 float SysCur = 0;
-float Throttle = 0;
+int Throttle = 0;
 
 /* ADC data offset */
 float ACCur_1_Offset = 0.0f;
@@ -91,6 +91,6 @@ void getMeasurement()
 	H1 = (float)ADC_DMA_array[1];		// H1
 	H3 = (float)ADC_DMA_array[4];		// H3
 	H2 = (float)ADC_DMA_array[5];		// H2
-	Throttle = (float)ADC_DMA_array[2];		// Throttle
+	Throttle = ADC_DMA_array[2];		// Throttle
 
 }
